Use explicit cast to make VS happy

+++ db/version_set.cc
@@ -2945,8 +2945,8 @@ void VersionStorageInfo::ComputeFilesMarkedForForcedBlobGC(
   }
 
   // Number of blob files eligible for GC based on age
-  const size_t cutoff_count =
-      blob_garbage_collection_age_cutoff * blob_files_.size();
+  const size_t cutoff_count = static_cast<size_t>(
+      blob_garbage_collection_age_cutoff * blob_files_.size());
   if (!cutoff_count) {
     return;
   }
